+++ flintrock/services.py
@@ -1,0 +1,336 @@
+import json
+import os
+import shlex
+import sys
+import textwrap
+import urllib.request
+
+# External modules
+import paramiko
+
+# Flintrock modules
+from .core import FlintrockCluster
+from .core import ssh_check_output
+
+THIS_DIR = os.path.dirname(os.path.realpath(__file__))
+
+
+# TODO: Cache these files. (?) They are being read potentially tens or
+#       hundreds of times. Maybe it doesn't matter because the files
+#       are so small.
+def get_formatted_template(path: str, mapping: dict) -> str:
+    with open(path) as f:
+        formatted = f.read().format(**mapping)
+    return formatted
+
+
+class FlintrockService:
+    """
+    This is an abstract class. Implementations of this class capture all the logic
+    required to fully install and manage services like Spark on Flintrock clusters.
+    """
+
+    def __init__(self):
+        """
+        This is the only method signature that implementations don't have to follow.
+        Use this method to set properties like the service version or download source
+        which the rest of the methods here will need to do their work.
+        """
+        raise NotImplementedError
+
+    def install(
+            self,
+            ssh_client: paramiko.client.SSHClient,
+            cluster: FlintrockCluster):
+        """
+        Install the service on a node via the provided SSH client. This typically
+        means downloading a software package and maybe even building it if necessary.
+
+        This method is role-agnostic; it runs on both the cluster master and slaves.
+        This method is meant to be called asynchronously.
+        """
+        raise NotImplementedError
+
+    def configure(
+            self,
+            ssh_client: paramiko.client.SSHClient,
+            cluster: FlintrockCluster):
+        """
+        Configure the installed service on a node via the provided SSH client. This
+        typically means using templates to create configuration files on the node.
+
+        This method is role-agnostic; it runs on both the cluster master and slaves.
+        This method is meant to be called asynchronously.
+        """
+        raise NotImplementedError
+
+    def configure_master(
+            self,
+            ssh_client: paramiko.client.SSHClient,
+            cluster: FlintrockCluster):
+        """
+        Configure the service master on a node via the provided SSH client after the
+        role-agnostic configuration in configure() is complete. Start the master and
+        slaves.
+
+        This method is meant to be called once on the cluster master.
+        This method is meant to be called asynchronously.
+        """
+        raise NotImplementedError
+
+    def configure_slave(
+            self,
+            ssh_client: paramiko.client.SSHClient,
+            cluster: FlintrockCluster):
+        """
+        Configure a service slave on a node via the provided SSH client after the
+        role-agnostic configuration in configure() is complete.
+
+        This method is meant to be called once on each cluster slave.
+        This method is meant to be called asynchronously.
+        """
+        raise NotImplementedError
+
+    def health_check(
+            self,
+            master_host: str):
+        """
+        Check that the service is up and running by querying the cluster master.
+        """
+        raise NotImplementedError
+
+
+class HDFS(FlintrockService):
+    def __init__(self, version):
+        self.version = version
+
+    def install(
+            self,
+            ssh_client: paramiko.client.SSHClient,
+            cluster: FlintrockCluster):
+        print("[{h}] Installing HDFS...".format(
+            h=ssh_client.get_transport().getpeername()[0]))
+
+        with ssh_client.open_sftp() as sftp:
+            sftp.put(
+                localpath=os.path.join(THIS_DIR, 'download-hadoop.py'),
+                remotepath='/tmp/download-hadoop.py')
+
+        ssh_check_output(
+            client=ssh_client,
+            command="""
+                set -e
+
+                python /tmp/download-hadoop.py "{version}"
+
+                mkdir "hadoop"
+                mkdir "hadoop/conf"
+
+                tar xzf "hadoop-{version}.tar.gz" -C "hadoop" --strip-components=1
+                rm "hadoop-{version}.tar.gz"
+            """.format(version=self.version))
+
+    def configure(
+            self,
+            ssh_client: paramiko.client.SSHClient,
+            cluster: FlintrockCluster):
+        # TODO: os.walk() through these files.
+        template_paths = [
+            'hadoop/conf/masters',
+            'hadoop/conf/slaves',
+            'hadoop/conf/hadoop-env.sh',
+            'hadoop/conf/core-site.xml',
+            'hadoop/conf/hdfs-site.xml']
+
+        for template_path in template_paths:
+            ssh_check_output(
+                client=ssh_client,
+                command="""
+                    echo {f} > {p}
+                """.format(
+                    f=shlex.quote(
+                        get_formatted_template(
+                            path=os.path.join(THIS_DIR, "templates", template_path),
+                            mapping=cluster.generate_template_mapping(service='hdfs'))),
+                    p=shlex.quote(template_path)))
+
+    # TODO: Convert this into start_master() and split master- or slave-specific
+    #       stuff out of configure() into configure_master() and configure_slave().
+    def configure_master(
+            self,
+            ssh_client: paramiko.client.SSHClient,
+            cluster: FlintrockCluster):
+        host = ssh_client.get_transport().getpeername()[0]
+        print("[{h}] Configuring HDFS master...".format(h=host))
+
+        ssh_check_output(
+            client=ssh_client,
+            command="""
+                ./hadoop/bin/hdfs namenode -format -nonInteractive
+                ./hadoop/sbin/start-dfs.sh
+            """)
+
+    def health_check(self, master_host: str):
+        # This info is not helpful as a detailed health check, but it gives us
+        # an up / not up signal.
+        hdfs_master_ui = 'http://{m}:50070/webhdfs/v1/?op=GETCONTENTSUMMARY'.format(m=master_host)
+
+        try:
+            hdfs_ui_info = json.loads(
+                urllib.request.urlopen(hdfs_master_ui).read().decode('utf-8'))
+        except Exception as e:
+            # TODO: Catch a more specific problem.
+            # TODO: Don't print to screen here. Raise the exception and let the
+            #       catcher decide on whether to print.
+            print("HDFS health check failed.", file=sys.stderr)
+            raise
+
+        print("HDFS online.")
+
+
+class Spark(FlintrockService):
+    def __init__(self, version: str=None, git_commit: str=None, git_repository: str=None):
+        # TODO: Convert these checks into something that throws a proper exception.
+        #       Perhaps reuse logic from CLI.
+        assert bool(version) ^ bool(git_commit)
+        if git_commit:
+            assert git_repository
+
+        self.version = version
+        self.git_commit = git_commit
+        self.git_repository = git_repository
+
+    def install(
+            self,
+            ssh_client: paramiko.client.SSHClient,
+            cluster: FlintrockCluster):
+        # TODO: Allow users to specify the Spark "distribution". (?)
+        distribution = 'hadoop2.6'
+
+        print("[{h}] Installing Spark...".format(
+            h=ssh_client.get_transport().getpeername()[0]))
+
+        try:
+            if self.version:
+                with ssh_client.open_sftp() as sftp:
+                    sftp.put(
+                        localpath=os.path.join(THIS_DIR, 'install-spark.sh'),
+                        remotepath='/tmp/install-spark.sh')
+                    sftp.chmod(path='/tmp/install-spark.sh', mode=0o755)
+                ssh_check_output(
+                    client=ssh_client,
+                    command="""
+                        set -e
+                        /tmp/install-spark.sh {spark_version} {distribution}
+                        rm -f /tmp/install-spark.sh
+                    """.format(
+                            spark_version=shlex.quote(self.version),
+                            distribution=shlex.quote(distribution)))
+            else:
+                ssh_check_output(
+                    client=ssh_client,
+                    command="""
+                        set -e
+                        sudo yum install -y git
+                        sudo yum install -y java-devel
+                        """)
+                ssh_check_output(
+                    client=ssh_client,
+                    command="""
+                        set -e
+                        git clone {repo} spark
+                        cd spark
+                        git reset --hard {commit}
+                        ./make-distribution.sh -T 1C -Phadoop-2.6
+                    """.format(
+                        repo=shlex.quote(self.git_repository),
+                        commit=shlex.quote(self.git_commit)))
+        except Exception as e:
+            # TODO: This should be a more specific exception.
+            print("Error: Failed to install Spark.", file=sys.stderr)
+            print(e, file=sys.stderr)
+            raise
+
+    def configure(
+            self,
+            ssh_client: paramiko.client.SSHClient,
+            cluster: FlintrockCluster):
+        template_paths = [
+            'spark/conf/spark-env.sh',
+            'spark/conf/slaves']
+        for template_path in template_paths:
+            ssh_check_output(
+                client=ssh_client,
+                command="""
+                    echo {f} > {p}
+                """.format(
+                    f=shlex.quote(
+                        get_formatted_template(
+                            path=os.path.join(THIS_DIR, "templates", template_path),
+                            mapping=cluster.generate_template_mapping(service='spark'))),
+                    p=shlex.quote(template_path)))
+
+    # TODO: Convert this into start_master() and split master- or slave-specific
+    #       stuff out of configure() into configure_master() and configure_slave().
+    #       start_slave() can block until slave is fully up; that way we don't need
+    #       a sleep() before starting the master.
+    def configure_master(
+            self,
+            ssh_client: paramiko.client.SSHClient,
+            cluster: FlintrockCluster):
+        host = ssh_client.get_transport().getpeername()[0]
+        print("[{h}] Configuring Spark master...".format(h=host))
+
+        # TODO: Maybe move this shell script out to some separate file/folder
+        #       for the Spark service.
+        # TODO: Add some timeout for waiting on master UI to come up.
+        ssh_check_output(
+            client=ssh_client,
+            command="""
+                set -e
+
+                spark/sbin/start-master.sh
+
+                set +e
+
+                master_ui_response_code=0
+                while [ "$master_ui_response_code" -ne 200 ]; do
+                    sleep 1
+                    master_ui_response_code="$(
+                        curl --head --silent --output /dev/null \
+                             --write-out "%{{http_code}}" {m}:8080
+                    )"
+                done
+
+                set -e
+
+                spark/sbin/start-slaves.sh
+            """.format(
+                m=shlex.quote(cluster.master_host)))
+
+    def health_check(self, master_host: str):
+        spark_master_ui = 'http://{m}:8080/json/'.format(m=master_host)
+
+        try:
+            spark_ui_info = json.loads(
+                urllib.request.urlopen(spark_master_ui).read().decode('utf-8'))
+        except Exception as e:
+            # TODO: Catch a more specific problem known to be related to Spark not
+            #       being up; provide a slightly better error message, and don't
+            #       dump a large stack trace on the user.
+            print("Spark health check failed.", file=sys.stderr)
+            raise
+
+        # TODO: Don't print here. Return this and let the caller print.
+        print(textwrap.dedent(
+            """\
+            Spark Health Report:
+              * Master: {status}
+              * Workers: {workers}
+              * Cores: {cores}
+              * Memory: {memory:.1f} GB\
+            """.format(
+                status=spark_ui_info['status'],
+                workers=len(spark_ui_info['workers']),
+                cores=spark_ui_info['cores'],
+                memory=spark_ui_info['memory'] / 1024)))
